requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.3
cron-descriptor==1.4.3
Django==5.0.2
django-celery-beat==2.6.0
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.35.2
greenlet==3.0.3
idna==3.6
inflection==0.5.1
kombu==5.3.5
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
stripe==8.7.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
